sys_locale::unix

Function _get

fn _get(env: &impl EnvAccess) -> impl Iterator<Item = String>

Retrieves a list of unique locales by checking specific environment variables in a predefined order: LANGUAGE, LC_ALL, LC_MESSAGES, and LANG.

The function first checks the LANGUAGE environment variable, which can contain one or more locales separated by a colon (:). It then splits these values, converts them from POSIX to BCP 47 format, and adds them to the list of locales if they are not already included.

Next, the function checks the LC_ALL, LC_MESSAGES, and LANG environment variables. Each of these variables contains a single locale. If a locale is found, and it’s not empty, it is converted to BCP 47 format and added to the list if it is not already included.

For more information check this issue: https://github.com/1Password/sys-locale/issues/14.

The function ensures that locales are returned in the order of precedence and without duplicates. The final list of locales is returned as an iterator.

Returns

An iterator over the unique locales found in the environment variables.

Environment Variables Checked

LANGUAGE - Can contain multiple locales, each separated by a colon (:), highest priority.

LC_ALL - Contains a single locale, high priority.

LC_MESSAGES - Contains a single locale, medium priority.

LANG - Contains a single locale, low priority.

Example

let locales: Vec<String> = _get(&env).collect();
for locale in locales {
    println!("User's preferred locales: {}", locale);
}