Crate sys_locale

A library to safely and easily obtain the current locale on the system or for an application.

This library currently supports the following platforms:

Android

iOS

macOS

Linux, BSD, and other UNIX variations

WebAssembly on the web (via the js feature)

Windows

Functions

get_locale

Returns the active locale for the system or application.

get_locales

Returns the preferred locales for the system or application, in descending order of preference.